feat: nodeVersionManager
