+++ renderer/src/pages/Node/components/InstallStep/index.module.scss
@@ -14,7 +14,7 @@
         .next-step-item-node-placeholder {
           width: 20px;
           height: 20px;
-  
+
           .next-step-item-node {
             .next-icon:before {
               width: 20px;
